be more careful about nuking volumes in reset script

+++ resetDocker.sh
@@ -1,5 +1,23 @@
+#!/usr/bin/env bash
+
+VOLUMES=0
+while [[ $# -gt 0 ]]; do
+  key="$1"
+  case $key in
+    --volumes)
+      VOLUMES=1
+      shift # past argument
+      ;;
+    *)    # unknown option
+      shift # past argument
+      ;;
+  esac
+done
+
 docker-compose down
 docker system prune -f
-docker system prune --volumes -f
+if (($VOLUMES)); then
+  docker system prune --volumes -f
+fi
 docker-compose build
 docker-compose up
